requirements.txt
playwright
beautifulsoup4
readability-lxml
azure-cognitiveservices-speech
python-dotenv
asyncio
